% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_clustering.R
\name{consensus_clustering}
\alias{consensus_clustering}
\title{Clustering of TIPC metrics}
\usage{
consensus_clustering(
  min_k = 2,
  max_k = 6,
  distance = "Pearson",
  root_dir = NULL,
  output_bnm = "test",
  seed = 999
)
}
\arguments{
\item{min_k}{An integer indicating the minimum number of clusters.}

\item{max_k}{An integer indicating the maximum number of clusters.}

\item{distance}{A character value indicating the distance method used in
\code{\link[ConsensusClusterPlus]{ConsensusClusterPlus}}}

\item{root_dir}{A directory path containing the TIPC_metrics, i.e. normalized
TIPC metrics of all 5 directions output from
\code{\link[TIPC]{normalize_metrics}}; nrow = total. of tumors, ncol = 6
(TIPC metrics) x 5 (directions).}

\item{output_bnm}{A character string appended to output folder name;
sub-folders are created for different k from min_k to max_k.}

\item{seed}{An integer seed for randomizing case ordering and for ConsensusClusterPlus.}
}
\description{
Clustering of TIPC metrics using
\code{\link[ConsensusClusterPlus]{ConsensusClusterPlus}}
}
\examples{
root_dir <- system.file("data", package = "TIPC")
consensus_clustering(root_dir=root_dir)
}
