% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_TIPC_cat.R
\name{count_TIPC_cat}
\alias{count_TIPC_cat}
\title{TIPC spatial metric counting}
\usage{
count_TIPC_cat(root_dir = NULL)
}
\arguments{
\item{root_dir}{A directory path to the tessellated_list, i.e. A list of
length (6x5): 6 TIPC categories & 5 directions; which is output from
\code{\link[TIPC]{tessellation}}.}
}
\value{
A list contain all 5 directional TIPC metrics which will be saved in
  \code{root_dir}
}
\description{
Count number of sub-regsions of each of the TIPC spatial category, for
individual directions
}
