% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{density_data}
\alias{density_data}
\title{Tumor denisty data.}
\format{A data.frame with 50 rows (tumor_ids in row.names) and 2 variables:
\describe{
  \item{density_ct1}{density values of celltype1}
  \item{density_ct2}{density values of celltype2}
  }}
\usage{
density_data
}
\description{
A data.frame containing containing 50 rows (i.e. tumors) and 2 variables
}
\keyword{datasets}
