% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_hexLen_count_TIPC_cat.R
\name{multiple_hexLen_count_TIPC_cat}
\alias{multiple_hexLen_count_TIPC_cat}
\title{A wrapper function of \code{\link[TIPC]{count_TIPC_cat}}}
\usage{
multiple_hexLen_count_TIPC_cat(
  root_dir = NULL,
  hex_len_range = NULL,
  hex_len_stepsize = 10
)
}
\arguments{
\item{root_dir}{A directory path pointing to the parent directory of all the
ouptput folders generated by \code{\link[TIPC]{tessellation}} at various
hexagonal lengths.}

\item{hex_len_range}{A vector of 2 integer number indicating the range of
hexagonal length. If unspecified (NULL), all found result folders generated
by \code{\link[TIPC]{tessellation}} will be processed.}

\item{hex_len_stepsize}{An integer indicating the step size of hexagonal
length. If unspecified (NULL), all found result folders generated
by \code{\link[TIPC]{tessellation}} will be processed.}
}
\description{
A wrapper function to run \code{\link[TIPC]{count_TIPC_cat}} at various
hexagonal lengths.
}
