% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_hexLen_tessellation.R
\name{multiple_hexLen_tessellation}
\alias{multiple_hexLen_tessellation}
\title{A wrapper function of \code{\link[TIPC]{tessellation}}}
\usage{
multiple_hexLen_tessellation(
  cell_data = NULL,
  output_dir = NULL,
  hex_len_range = c(40, 60),
  hex_len_stepsize = 10
)
}
\arguments{
\item{cell_data}{A data.frame of cell data of size rxc; r = total cells
across all tumors and cores, 5 columns named as 'Phenotype', 'X', 'Y',
'tumor_ids', 'core_ids'. NOTE. 'tumor_ids' & 'core_ids' contain no-space
and no underscore}

\item{output_dir}{A directory path for results saving.}

\item{hex_len_range}{A vector of 2 integer number indicating the range of hexagonal length.}

\item{hex_len_stepsize}{An integer indicating the step size of hexagonal length.}
}
\value{
A list contain all 5 directional sub-region count of 6 TIPC category.
}
\description{
A wrapper function to run \code{\link[TIPC]{tessellation}} at various hexagonal lengths.
}
