% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_metrics.R
\name{normalize_metrics}
\alias{normalize_metrics}
\title{Normalization of TIPC spatial metrics}
\usage{
normalize_metrics(root_dir = NULL)
}
\arguments{
\item{root_dir}{A directory path containing the TIPC_counts, i.e. counts of
TIPC categories of 5 shifted grids output from
\code{\link[TIPC]{count_TIPC_cat}}.}
}
\description{
Calculate normalized TIPC spatial metrics for individual directions.
}
