% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_hexLen.r
\name{optimal_hexLen}
\alias{optimal_hexLen}
\title{Determine the optimal hexagon length}
\usage{
optimal_hexLen(data, output_dir)
}
\arguments{
\item{data}{A data.frame containing the 6 normalized TIPC spatial parameters (in %)
in all 5 directions across all hexLen output from
\code{\link[TIPC]{trend_plot_hexLen}}.}

\item{output_dir}{A character string indicating the output folder name; the directory the output pdf will be saved.}
}
\value{
A character indicating the optimal hexLen.
}
\description{
Determine the optimal hexagon length based on
(i) the smallest difference in Q3-Q1 across the 6 TIPC spatial parameters, and
(ii) the smallest variance/sd across 5 directions within each TIPC spatial parameters
}
