% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_k.r
\name{optimal_k}
\alias{optimal_k}
\title{Determine the optimal cluster number k}
\usage{
optimal_k(optHex_dir)
}
\arguments{
\item{optHex_dir}{A character string indicating the folder name corresponding the selected optimal hexLen
which can be determined using \code{\link[TIPC]{optimal_hexLen}}.}

\item{data}{A data.frame containing the 6 normalized TIPC spatial parameters (in %)
in all 5 directions across all hexLen output from
\code{\link[TIPC]{trend_plot_hexLen}}.}
}
\value{
A character indicating the optimal cluster k.
}
\description{
Determine the optimal cluster number k based on 
(i) first finding the shoulder point in cumulative CDF plot representing the smallest k with stability, and followed up
(ii) finding the largest (most granular) k in tracking plot that gives the highest information.
}
