% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TIPC_heatmap_overlayDens.R
\name{plot_TIPC_heatmap_overlayDens}
\alias{plot_TIPC_heatmap_overlayDens}
\title{Heat-map of TIPC metric color-coded by clustering results}
\usage{
plot_TIPC_heatmap_overlayDens(
  root_dir = NULL,
  one_k = NULL,
  clustering_subfolder_nm = NULL,
  density_data = NULL
)
}
\arguments{
\item{root_dir}{A directory path containing (1) the TIPC_metrics, i.e.
normalized TIPC metrics of all 5 directions output from
\code{\link[TIPC]{normalize_metrics}}; (2) the result folder of TIPC
clustering results output from \code{\link[TIPC]{consensus_clustering}}.}

\item{one_k}{An integer specifying the cluster number selected for heat-map
plotting; otherwise all k's found under \code{TIPC_cluster_dir} will be
processed.}

\item{clustering_subfolder_nm}{A character string of sub-folder name generated
during \code{\link[TIPC]{consensus_clustering}}, parked under \code{root_dir}.}

\item{density_data}{A data.frame containing the tumor-level density data of
no. of rows equals to no. of cases/tumors (tumor_ids in row.names), each column
represents the density values of a specific cell-type.}
}
\description{
Plotting heat-maps of TIPC metric clustering results of all k's unless it is
specified; using R package \code{\link[ComplexHeatmap]{Heatmap}}
}
\examples{
root_dir <- system.file("data", package = "TIPC")
clustering_subfolder_nm <- 'ConsensusClusterPlus_test'
plot_TIPC_heatmap(root_dir = root_dir, clustering_subfolder_nm = clustering_subfolder_nm)
}
