% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postTIPC_survival_analysis.R
\name{postTIPC_SurvivalAnalysis}
\alias{postTIPC_SurvivalAnalysis}
\title{Heat-map of TIPC metric color-coded by clustering results}
\usage{
postTIPC_SurvivalAnalysis(
  root_dir = NULL,
  clustering_subfolder_nm = NULL,
  one_k = NULL,
  min_cluster_size = 30,
  method = c("univariate", "multivariate"),
  surv_data = NULL,
  ref_cluster_no = NULL,
  all_ref_check = FALSE,
  KM_xlab = "Time",
  bnm = ""
)
}
\arguments{
\item{root_dir}{A directory path containing (1) the TIPC_metrics, i.e.
normalized TIPC metrics of all 5 directions output from
\code{\link[TIPC]{normalize_metrics}}; (2) the result folder of TIPC
clustering results output from \code{\link[TIPC]{consensus_clustering}}.}

\item{clustering_subfolder_nm}{A character string of sub-folder name
generated during \code{\link[TIPC]{consensus_clustering}}, parked under
\code{root_dir}.}

\item{one_k}{An integer specifying the cluster number selected for heat-map
plotting; otherwise all k's found under \code{TIPC_cluster_dir} will be
processed.}

\item{min_cluster_size}{An integer indicating the minimum size of TIPC
cluster to be tested for survival analysis; default 10.}

\item{method}{A character string indicating the method of survival analysis:
univariate or multivariate.}

\item{surv_data}{A data frame of size r rows x 2 (or more) columns:
containing r tumors and minimum 2 variables with column names 'time'
(numeric) and 'cens' (integer), any additional columns are used as
covariates (factors) if multivariate analysis is chosen. NOTE: rownames
containing tumor_ids.}

\item{ref_cluster_no}{An integer indicating the cluster id to be used as the
reference, if unspecified (NULL) cluster of largest size will be used.}

\item{all_ref_check}{A boolean indicating if multiple survival analyses are performed
using individual clusters as reference; if TRUE, it will overwrite \code{ref_cluster_no}}

\item{KM_xlab}{A character string specifying the x-axis label of Kaplan-Meier curve}

\item{bnm}{A character string appending to the output filenames}
}
\description{
Plotting heat-maps of TIPC metric clustering results of all k's unless it is
specified; using R package \code{\link[survival]{coxph}}
}
