% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surv_data}
\alias{surv_data}
\title{Clinical outcome and clinicopathological data.}
\format{
A data frame containing 50 rows (i.e. tumors) and 16 variables:
  \describe{ \item{time}{survival time, numeric} \item{cens}{censoring data,
  binary} \item{covariate_1}{covariate var1, factor}
  \item{covariate_2}{covariate var, factor} \item{covariate_3}{covariate var, factor}
  \item{covariate_4}{covariate var, factor} \item{covariate_5}{covariate var, factor}
  \item{covariate_6}{covariate var, factor} \item{covariate_7}{covariate var, factor}
  \item{covariate_8}{covariate var, factor} \item{covariate_9}{covariate var1 factor}
  \item{covariate_10}{covariate var, factor} \item{covariate_11}{covariate var1 factor}
  \item{covariate_12}{covariate var, factor} \item{covariate_13}{covariate var, factor}
  \item{covariate_14}{covariate var, factor}}
}
\usage{
surv_data
}
\description{
A data frame containing the survival and clinicopathological data for 50 tumors.
}
\keyword{datasets}
