% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tessellated_res}
\alias{tessellated_res}
\title{Tesselated data.}
\format{
A list containing 3 (celltypes) x 5 (directions) elements; each
  element is a list of length 137 (cores) each of which is a named quadratcount
  table generated using \code{\link[spatstat.geom]{quadratcount}}.
}
\usage{
tessellated_res
}
\description{
A list containing cell counts in the 5-directional tesselated sub-regions
i.e. hexagons for immune cells, stromal cells and tumor cells; using hex_len
= 100.
}
\keyword{datasets}
