% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessellation.R
\name{tessellation}
\alias{tessellation}
\title{Cell space tessellation into hexagonal sub-regions}
\usage{
tessellation(
  cell_data = NULL,
  hex_len = 100,
  output_dir = NULL,
  plot_flag = FALSE
)
}
\arguments{
\item{cell_data}{A data.frame of cell data of size rxc; r = total cells
across all tumors and cores, 5 columns named as 'Phenotype', 'X', 'Y',
'tumor_ids', 'core_ids'. NOTE. 'tumor_ids' & 'core_ids' contain no-space
and no underscore}

\item{hex_len}{An integer of hexagonal length.}

\item{output_dir}{A directory path for results saving.}

\item{plot_flag}{A boolean for plotting option of cell image.}
}
\value{
A list contain all 5 directional sub-region count of 6 TIPC category
}
\description{
Sub-dividing 2D cell space into hexogonal sub-regions and counting the
number of immune cell, tumor cell, and stromal cell in each sub-region.
}
