% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIPC-package.R
\docType{package}
\name{TIPC-package}
\alias{TIPC}
\alias{TIPC-package}
\title{An R package for Tumor-immune Partitioning and Clustering (TIPC)}
\description{
An R package for Tumor-immune Partitioning and Clustering (TIPC), a computational algorithm for discovering and documenting spatially-derived signatures of tumor-immune cell interactions within the tumor microenvironment.
}
\author{
\strong{Maintainer}: MaiChan Lau \email{mclau@bwh.harvard.edu}

Authors:
\itemize{
  \item Jennifer Borowsky \email{borowsky.jennifer@gmail.com}
  \item Shuji Ogino \email{SOGINO@BWH.HARVARD.EDU}
  \item Jonathan Nowak \email{janowak@bwh.harvard.edu}
}

}
\keyword{internal}
