% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TIPC_counts}
\alias{TIPC_counts}
\title{TIPC count data.}
\format{A list containing 5 (directions) elements; each of which is a
  data.frame with 137 rows (i.e. cores) and 10 variables: \describe{
  \item{tumor_ONLY}{number of sub-region containing tumor cell only, counts}
  \item{Stroma_ONLY}{number of sub-region containing stromal cell only,
  counts} \item{I2Tu_high}{number of sub-region containing immune & tumor
  cells and I:T ratio > global I:T, counts} \item{I2Tu_low}{number of
  sub-region containing immune & tumor cells and I:T ratio < global I:T,
  counts} \item{I2S_high}{number of sub-region containing immune & tumor
  cells and I:S ratio > global I:S, counts} \item{I2S_low}{number of
  sub-region containing immune & tumor cells and I:S ratio < global I:S,
  counts} \item{empty}{number of sub-region 0 cell, counts}
  \item{global_I2Tu}{global I:T ratio, numeric} \item{global_I2S}{global I:S
  ratio, numeric} \item{total_subregions}{number of total sub-regions, counts}}}
\usage{
TIPC_counts
}
\description{
A list containing the counts of TIPC classes of the 5 directional grids.
}
\keyword{datasets}
