% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TIPC_metrics}
\alias{TIPC_metrics}
\title{TIPC normalized metric data.}
\format{
A data frame containing 50 rows (i.e. tumors) and 30 variables (6
  TIPC metrics x 5 directions): \describe{
  \item{tumor_ONLY_O}{fractions of sub-region containing tumor cell only, center}
  \item{Stroma_ONLY_O}{fractions of sub-region containing stromal cell only,
  center} \item{I2Tu_high_O}{fractions of sub-region containing immune & tumor
  cells and I:T ratio > global I:T, center} \item{I2Tu_low_O}{fractions of
  sub-region containing immune & tumor cells and I:T ratio < global I:T,
  center} \item{I2S_high_O}{fractions of sub-region containing immune & tumor
  cells and I:S ratio > global I:S, center} \item{I2S_low_O}{fractions of
  sub-region containing immune & tumor cells and I:S ratio < global I:S,
  center} \item{tumor_ONLY_L}{fractions of sub-region containing tumor cell only, left}
  \item{Stroma_ONLY_L}{fractions of sub-region containing stromal cell only,
  left} \item{I2Tu_high_L}{fractions of sub-region containing immune & tumor
  cells and I:T ratio > global I:T, left} \item{I2Tu_low_L}{fractions of
  sub-region containing immune & tumor cells and I:T ratio < global I:T,
  left} \item{I2S_high_L}{fractions of sub-region containing immune & tumor
  cells and I:S ratio > global I:S, left} \item{I2S_low_L}{fractions of
  sub-region containing immune & tumor cells and I:S ratio < global I:S,
  left} \item{tumor_ONLY_R}{fractions of sub-region containing tumor cell only, right}
  \item{Stroma_ONLY_R}{fractions of sub-region containing stromal cell only,
  right} \item{I2Tu_high_R}{fractions of sub-region containing immune & tumor
  cells and I:T ratio > global I:T, right} \item{I2Tu_low_R}{fractions of
  sub-region containing immune & tumor cells and I:T ratio < global I:T,
  right} \item{I2S_high_R}{fractions of sub-region containing immune & tumor
  cells and I:S ratio > global I:S, right} \item{I2S_low_R}{fractions of
  sub-region containing immune & tumor cells and I:S ratio < global I:S,
  right}\item{tumor_ONLY_U}{fractions of sub-region containing tumor cell only, upper}
  \item{Stroma_ONLY_U}{fractions of sub-region containing stromal cell only,
  upper} \item{I2Tu_high_U}{fractions of sub-region containing immune & tumor
  cells and I:T ratio > global I:T, upper} \item{I2Tu_low_U}{fractions of
  sub-region containing immune & tumor cells and I:T ratio < global I:T,
  upper} \item{I2S_high_U}{fractions of sub-region containing immune & tumor
  cells and I:S ratio > global I:S, upper} \item{I2S_low_U}{fractions of
  sub-region containing immune & tumor cells and I:S ratio < global I:S,
  upper}\item{tumor_ONLY_D}{fractions of sub-region containing tumor cell only, lower}
  \item{Stroma_ONLY_D}{fractions of sub-region containing stromal cell only,
  lower} \item{I2Tu_high_D}{fractions of sub-region containing immune & tumor
  cells and I:T ratio > global I:T, lower} \item{I2Tu_low_D}{fractions of
  sub-region containing immune & tumor cells and I:T ratio < global I:T,
  lower} \item{I2S_high_D}{fractions of sub-region containing immune & tumor
  cells and I:S ratio > global I:S, lower} \item{I2S_low_D}{fractions of
  sub-region containing immune & tumor cells and I:S ratio < global I:S,
  lower}}
}
\usage{
TIPC_metrics
}
\description{
A data frame containing the normalized TIPC counts with all 5 directions
concatenated.
}
\keyword{datasets}
