% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_plot_shiftDirection.r
\name{trend_plot_shiftDirection}
\alias{trend_plot_shiftDirection}
\title{Trend plot of grid shifting effect}
\usage{
trend_plot_shiftDirection(
  root_dir = NULL,
  hex_len_range = NULL,
  hex_len_stepsize = NULL,
  pdf_width = 9,
  pdf_height = 12
)
}
\arguments{
\item{root_dir}{A directory path pointing to the parent directory of the
ouptput folders containing TIPC metric counts generated by
\code{\link[TIPC]{count_TIPC_cat}} at various hexagonal lengths.}

\item{hex_len_range}{A vector of 2 integer number indicating the range of
hexagonal length. If unspecified (NULL), all found TIPC result folders of
different hexagonal lengths will be processed. All hex_len data will
be collapsed into 5 directions.}

\item{hex_len_stepsize}{An integer indicating the step size of hexagonal
length. If unspecified (NULL), all found TIPC result folders of
different hexagonal lengths will be processed.}

\item{pdf_width, pdf_height}{the width and height of the graphics region in inches.
The default values are 9 and 12; inherited from \code{\link[grDevices]{pdf}}}
}
\description{
Plotting the changes in TIPC metric count as the grid shfited to 4 other
directions i.e. origin + left + right + uppper + down
}
