% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_plot_tinyClusters.r
\name{trend_plot_tinyClusters}
\alias{trend_plot_tinyClusters}
\title{Trend plot of outlier clusteres}
\usage{
trend_plot_tinyClusters(
  root_dir = NULL,
  hex_len_range = NULL,
  hex_len_stepsize = NULL,
  pdf_width = 12,
  pdf_height = 8,
  min_cluster_size = 30,
  clustering_subfolder_nm = "ConsensusClusterPlus_test"
)
}
\arguments{
\item{root_dir}{A directory path pointing to the parent directory of the
ouptput folders containing TIPC metric counts generated by
\code{\link[TIPC]{count_TIPC_cat}} at various hexagonal lengths.}

\item{hex_len_range}{A vector of 2 integer number indicating the range of
hexagonal length. If unspecified (NULL), all found TIPC result folders of
different hexagonal lengths will be processed.}

\item{hex_len_stepsize}{An integer indicating the step size of hexagonal
length. If unspecified (NULL), all found TIPC result folders of
different hexagonal lengths will be processed.}

\item{pdf_width, pdf_height}{the width and height of the graphics region in inches.
The default values are 18 and 12; inherited from \code{\link[grDevices]{pdf}}}

\item{min_cluster_size}{An integer indicating the minimum size of TIPC
cluster to be tested for survival analysis; default 10.}

\item{clustering_subfolder_nm}{A character string indicating the subfolder
name containing the TIPC clustering results within each hex_len.}
}
\description{
Plotting the number of tiny clusters and unclassified cases
as the subregional size (hex_len) and different number of clusters (k) are
used.
}
\examples{
root_dir <- "C:/Users/Mai Chan Lau/Desktop/TIPC_package/test_run/"
root_dir <-'C:/Users/Mai Chan Lau/Desktop/multiplex_analysis/TIPC_package/test_run_mIF_Tcell_bybatch30'
trend_plot_tinyClusters(root_dir = root_dir)

}
