% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_data}
\alias{cell_data}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 283015 rows and 5 variables:
\describe{
  \item{Phenotype}{cell types, i.e. tumor, stroma, immune}
  \item{X}{X coordinate of cells, in pixels}
  \item{Y}{Y coordinate of cells, in pixels}
  \item{tumor_ids}{tumor or patient identifiers, character strings}
  \item{core_ids}{core identifiers, character strings}
}}
\usage{
cell_data
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
