% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_clustering.R
\name{consensus_clustering}
\alias{consensus_clustering}
\title{Clustering of TIPC metrics}
\usage{
consensus_clustering(min_k = 2, max_k = 6, root_dir = NULL,
  output_bnm = "test")
}
\arguments{
\item{min_k}{An integer indicating the minimum number of clusters.}

\item{max_k}{An integer indicating the maximum number of clusters.}

\item{root_dir}{A directory path containing the TIPC_metrics, i.e. normalized
TIPC metrics of all 5 directions output from
\code{\link[TIPC]{normalize_metrics}}; nrow = total. of tumors, ncol = 6
(TIPC metrics) x 5 (directions).}

\item{output_bnm}{A character string appended to output folder name;
sub-folders are created for different k from min_k to max_k.}
}
\description{
Clustering of TIPC metrics using
\code{\link[ConsensusClusterPlus]{ConsensusClusterPlus}}
}
\examples{
root_dir <- "C:/Users/Mai Chan Lau/Desktop/TIPC_package/test_run/TIPC_hexLen100"
consensus_clustering(root_dir=root_dir)
}
