% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIPC-package.R
\docType{package}
\name{TIPC-package}
\alias{TIPC-package}
\title{TIPC: An R package for unsupervised characterization of immune cells in tumor microenvirontment.}
\description{
An R package for unsupervised characterization of immune cells in tumor microenvirontment.
}
\author{
\strong{Maintainer}: MaiChan Lau \email{mclau@bwh.harvard.edu}

Authors:
\itemize{
  \item Jonathan Nowak \email{janowak@bwh.harvard.edu}
  \item Shuji Ogino \email{SOGINO@BWH.HARVARD.EDU}
}

}
\keyword{internal}
